requirements.txt
streamlit>=1.28.0
watchdog>=3.0.0
pandas>=1.5.0
pillow>=9.0.0